requirements.txt
bidict==0.23.1
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-SocketIO==5.3.6
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
psutil==5.9.8
python-engineio==4.9.1
python-socketio==5.11.2
requests==2.32.3
simple-websocket==1.0.0
urllib3==2.2.1
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
